+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
-## [1.11.0] - ????-??-??
+## [1.11.0] - 2022-06-21
 
 ### Added